Removes Secondary & Final Objectives from Traitors
