+++ code/controllers/configuration/entries/game_options.dm
@@ -73,11 +73,6 @@
 	integer = FALSE
 	min_val = 0.01
 
-/// Determines how many potential objectives a traitor can have.
-/datum/config_entry/number/maximum_potential_objectives
-	default = 6
-	min_val = 1
-
 /datum/config_entry/number/changeling_scaling_coeff //how much does the amount of players get divided by to determine changelings
 	default = 6
 	integer = FALSE
