+++ code/modules/antagonists/traitor/datum_traitor.dm
@@ -23,8 +23,6 @@
 	var/uplink_flag_given = UPLINK_TRAITORS
 
 	var/give_objectives = TRUE
-	/// Whether to give secondary objectives to the traitor, which aren't necessary but can be completed for a progression and TC boost.
-	var/give_secondary_objectives = TRUE
 	var/should_give_codewords = TRUE
 	///give this traitor an uplink?
 	var/give_uplink = TRUE
@@ -53,17 +51,6 @@
 	///the final objective the traitor has to accomplish, be it escaping, hijacking, or just martyrdom.
 	var/datum/objective/ending_objective
 
-/datum/antagonist/traitor/infiltrator
-	// Used to denote traitors who have joined midround and therefore have no access to secondary objectives.
-	// Progression elements are best left to the roundstart antagonists
-	// There will still be a timelock on uplink items
-	name = "\improper Infiltrator"
-	give_secondary_objectives = FALSE
-	uplink_flag_given = UPLINK_INFILTRATORS
-
-/datum/antagonist/traitor/infiltrator/sleeper_agent
-	name = "\improper Syndicate Sleeper Agent"
-
 /datum/antagonist/traitor/New(give_objectives = TRUE)
 	. = ..()
 	src.give_objectives = give_objectives
@@ -86,10 +73,6 @@
 		uplink_handler.primary_objectives = objectives
 		uplink_handler.has_progression = TRUE
 		SStraitor.register_uplink_handler(uplink_handler)
-
-		if(give_secondary_objectives)
-			uplink_handler.has_objectives = TRUE
-			uplink_handler.generate_objectives()
 
 		uplink_handler.can_replace_objectives = CALLBACK(src, PROC_REF(can_change_objectives))
 		uplink_handler.replace_objectives = CALLBACK(src, PROC_REF(submit_player_objective))
@@ -120,61 +103,12 @@
 
 /datum/antagonist/traitor/on_removal()
 	if(!isnull(uplink_handler))
-		uplink_handler.has_objectives = FALSE
 		uplink_handler.can_replace_objectives = null
 		uplink_handler.replace_objectives = null
 	owner.take_uplink()
 	owner.special_role = null
 	owner.forget_crafting_recipe(/datum/crafting_recipe/syndicate_uplink_beacon)
 	return ..()
-
-/datum/antagonist/traitor/proc/traitor_objective_to_html(datum/traitor_objective/to_display)
-	var/string = "[to_display.name]"
-	if(to_display.objective_state == OBJECTIVE_STATE_ACTIVE || to_display.objective_state == OBJECTIVE_STATE_INACTIVE)
-		string += " <a href='byond://?src=[REF(owner)];edit_obj_tc=[REF(to_display)]'>[to_display.telecrystal_reward] TC</a>"
-		string += " <a href='byond://?src=[REF(owner)];edit_obj_pr=[REF(to_display)]'>[to_display.progression_reward] PR</a>"
-	else
-		string += ", [to_display.telecrystal_reward] TC"
-		string += ", [to_display.progression_reward] PR"
-	if(to_display.objective_state == OBJECTIVE_STATE_ACTIVE && !istype(to_display, /datum/traitor_objective/ultimate))
-		string += " <a href='byond://?src=[REF(owner)];fail_objective=[REF(to_display)]'>Fail this objective</a>"
-		string += " <a href='byond://?src=[REF(owner)];succeed_objective=[REF(to_display)]'>Succeed this objective</a>"
-	if(to_display.objective_state == OBJECTIVE_STATE_INACTIVE)
-		string += " <a href='byond://?src=[REF(owner)];fail_objective=[REF(to_display)]'>Dispose of this objective</a>"
-
-	if(to_display.skipped)
-		string += " - <b>Skipped</b>"
-	else if(to_display.objective_state == OBJECTIVE_STATE_FAILED)
-		string += " - <b><font color='red'>Failed</font></b>"
-	else if(to_display.objective_state == OBJECTIVE_STATE_INVALID)
-		string += " - <b>Invalidated</b>"
-	else if(to_display.objective_state == OBJECTIVE_STATE_COMPLETED)
-		string += " - <b><font color='green'>Succeeded</font></b>"
-
-	return string
-
-/datum/antagonist/traitor/antag_panel_objectives()
-	var/result = ..()
-	if(!uplink_handler)
-		return result
-	result += "<i><b>Traitor specific objectives</b></i><br>"
-	result += "<i><b>Concluded Objectives</b></i>:<br>"
-	for(var/datum/traitor_objective/objective as anything in uplink_handler.completed_objectives)
-		result += "[traitor_objective_to_html(objective)]<br>"
-	if(!length(uplink_handler.completed_objectives))
-		result += "EMPTY<br>"
-	result += "<i><b>Ongoing Objectives</b></i>:<br>"
-	for(var/datum/traitor_objective/objective as anything in uplink_handler.active_objectives)
-		result += "[traitor_objective_to_html(objective)]<br>"
-	if(!length(uplink_handler.active_objectives))
-		result += "EMPTY<br>"
-	result += "<i><b>Potential Objectives</b></i>:<br>"
-	for(var/datum/traitor_objective/objective as anything in uplink_handler.potential_objectives)
-		result += "[traitor_objective_to_html(objective)]<br>"
-	if(!length(uplink_handler.potential_objectives))
-		result += "EMPTY<br>"
-	result += "<a href='byond://?src=[REF(owner)];common=give_objective'>Force add objective</a><br>"
-	return result
 
 /// Returns true if we're allowed to assign ourselves a new objective
 /datum/antagonist/traitor/proc/can_change_objectives()
@@ -340,9 +274,6 @@
 				traitor_won = FALSE
 			objectives_text += "<br><B>Objective #[count]</B>: [objective.explanation_text] [objective.get_roundend_success_suffix()]"
 			count++
-		if(uplink_handler.final_objective)
-			objectives_text += "<br>[span_greentext("[traitor_won ? "Additionally" : "However"], the final objective \"[uplink_handler.final_objective]\" was completed!")]"
-			traitor_won = TRUE
 
 	result += "<br>[owner.name] <B>[traitor_flavor["roundend_report"]]</B>"
