+++ code/modules/events/stray_cargo.dm
@@ -166,7 +166,6 @@
 		"Nuke Op" = UPLINK_NUKE_OPS,
 		"Clown Op" = UPLINK_CLOWN_OPS,
 		"Lone Op" = UPLINK_LONE_OP,
-		"Infiltrator" = UPLINK_INFILTRATORS,
 		"Spy" = UPLINK_SPY
 		)
 	var/uplink_type = tgui_input_list(usr, "Choose uplink to draw items from.", "Choose uplink type.", possible_uplinks)
