+++ code/modules/paperwork/paper_cutter.dm
@@ -113,9 +113,12 @@
 /obj/item/papercutter/attackby(obj/item/inserted_item, mob/user, params)
 	if(istype(inserted_item, /obj/item/paper))
 		if(is_type_in_list(inserted_item, list(
-			/obj/item/paper/paperslip, /obj/item/paper/report, /obj/item/paper/fake_report,
-			/obj/item/paper/calling_card, /obj/item/paper/pamphlet, /obj/item/paper/holy_writ)
-			))
+				/obj/item/paper/fake_report,
+				/obj/item/paper/holy_writ,
+				/obj/item/paper/pamphlet,
+				/obj/item/paper/paperslip,
+				/obj/item/paper/report,
+		)))
 			balloon_alert(user, "won't fit!")
 			return
 		if(stored_paper)
