+++ code/modules/uplink/uplink_items/contractor.dm
@@ -13,7 +13,7 @@
 	item = /obj/item/storage/box/syndicate/contract_kit
 	category = /datum/uplink_category/contractor
 	cost = 20
-	purchasable_from = UPLINK_INFILTRATORS
+	purchasable_from = UPLINK_TRAITORS
 
 /datum/uplink_item/bundles_tc/contract_kit/purchase(mob/user, datum/uplink_handler/uplink_handler, atom/movable/source)
 	. = ..()
