+++ tgstation.dme
@@ -3360,40 +3360,13 @@
 #include "code\modules\antagonists\spy\spy_uplink.dm"
 #include "code\modules\antagonists\survivalist\survivalist.dm"
 #include "code\modules\antagonists\syndicate_monkey\syndicate_monkey.dm"
-#include "code\modules\antagonists\traitor\balance_helper.dm"
 #include "code\modules\antagonists\traitor\datum_traitor.dm"
-#include "code\modules\antagonists\traitor\objective_category.dm"
-#include "code\modules\antagonists\traitor\traitor_objective.dm"
 #include "code\modules\antagonists\traitor\uplink_handler.dm"
 #include "code\modules\antagonists\traitor\components\demoraliser.dm"
-#include "code\modules\antagonists\traitor\components\traitor_objective_helpers.dm"
-#include "code\modules\antagonists\traitor\components\traitor_objective_limit_per_time.dm"
-#include "code\modules\antagonists\traitor\components\traitor_objective_mind_tracker.dm"
 #include "code\modules\antagonists\traitor\contractor\contract_teammate.dm"
 #include "code\modules\antagonists\traitor\contractor\contractor_hub.dm"
 #include "code\modules\antagonists\traitor\contractor\contractor_items.dm"
 #include "code\modules\antagonists\traitor\contractor\syndicate_contract.dm"
-#include "code\modules\antagonists\traitor\objectives\assassination.dm"
-#include "code\modules\antagonists\traitor\objectives\demoralise_assault.dm"
-#include "code\modules\antagonists\traitor\objectives\destroy_heirloom.dm"
-#include "code\modules\antagonists\traitor\objectives\destroy_item.dm"
-#include "code\modules\antagonists\traitor\objectives\eyesnatching.dm"
-#include "code\modules\antagonists\traitor\objectives\hack_comm_console.dm"
-#include "code\modules\antagonists\traitor\objectives\infect.dm"
-#include "code\modules\antagonists\traitor\objectives\kidnapping.dm"
-#include "code\modules\antagonists\traitor\objectives\kill_pet.dm"
-#include "code\modules\antagonists\traitor\objectives\locate_weakpoint.dm"
-#include "code\modules\antagonists\traitor\objectives\sabotage_machinery.dm"
-#include "code\modules\antagonists\traitor\objectives\steal.dm"
-#include "code\modules\antagonists\traitor\objectives\abstract\target_player.dm"
-#include "code\modules\antagonists\traitor\objectives\final_objective\battle_royale.dm"
-#include "code\modules\antagonists\traitor\objectives\final_objective\battlecruiser.dm"
-#include "code\modules\antagonists\traitor\objectives\final_objective\final_objective.dm"
-#include "code\modules\antagonists\traitor\objectives\final_objective\infect_ai.dm"
-#include "code\modules\antagonists\traitor\objectives\final_objective\no_escape.dm"
-#include "code\modules\antagonists\traitor\objectives\final_objective\objective_dark_matteor.dm"
-#include "code\modules\antagonists\traitor\objectives\final_objective\romerol.dm"
-#include "code\modules\antagonists\traitor\objectives\final_objective\supermatter_cascade.dm"
 #include "code\modules\antagonists\valentines\heartbreaker.dm"
 #include "code\modules\antagonists\valentines\valentine.dm"
 #include "code\modules\antagonists\venus_human_trap\venus_human_trap.dm"
