--- tgui/packages/tgui/interfaces/Uplink/constants.ts
@@ -1,7 +1,0 @@
-export enum ObjectiveState {
-  Inactive = 1,
-  Active = 2,
-  Completed = 3,
-  Failed = 4,
-  Invalid = 5,
-}
